requirements.txt
# Dependencias principales
pyyaml==6.0.2
jsonschema==4.23.0
pyspark==3.5.1
pyarrow>=14.0.0
pandas>=2.0.0
requests>=2.31.0

# Dependencias para base de datos (Gold layer)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL

# Robustez (reintentos I/O)
tenacity==8.2.3

# Herramientas de estilo y calidad (desarrollo)
black==24.10.0
ruff==0.6.9
mypy==1.13.0
pre-commit==4.0.1
pytest==8.3.3
pytest-cov==4.1.0

# Dependencias opcionales (descomenta según necesidad)
# delta-spark==3.2.0
# great-expectations==0.18.12
# openlineage-python==1.27.0

# Utilidades adicionales
python-dotenv==1.0.0
